Moving into its own repo.
This was originally developed in the IETF NOC repo, but then I started using it in various other projects,
personal configs, etc. This was being done with symlinks, but I'd now like to start using it in another
project, and so it now needs to be more public.

It is expected that this will be used in all of these projects as a submodule.